+++ src/Model/FixerInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace YamlStandards\Model;
+
+use YamlStandards\Command\InputSettingData;
+
+interface FixerInterface
+{
+    /**
+     * @param string $pathToYamlFile
+     * @param string $pathToDumpFixedFile
+     * @param \YamlStandards\Command\InputSettingData $inputSettingData
+     * @return \YamlStandards\Result\Result
+     */
+    public function fix($pathToYamlFile, $pathToDumpFixedFile, InputSettingData $inputSettingData);
+}