+++ src/Model/YamlSpacesBetweenGroups/YamlSpacesBetweenGroupsChecker.php
@@ -1,10 +1,12 @@
 <?php
 
-namespace YamlStandards\Checker;
+namespace YamlStandards\Model\YamlSpacesBetweenGroups;
 
 use SebastianBergmann\Diff\Differ;
 use YamlStandards\Command\InputSettingData;
-use YamlStandards\Result;
+use YamlStandards\Command\ProcessOutput;
+use YamlStandards\Model\CheckerInterface;
+use YamlStandards\Result\Result;
 
 /**
  * Check yaml file have space between groups
@@ -28,13 +30,13 @@ class YamlSpacesBetweenGroupsChecker implements CheckerInterface
         }
 
         if ($yamlContent === $correctYamlContent) {
-            return new Result($pathToYamlFile, Result::RESULT_CODE_OK);
+            return new Result($pathToYamlFile, Result::RESULT_CODE_OK, ProcessOutput::STATUS_CODE_OK);
         }
 
         $differ = new Differ();
         $diffBetweenStrings = $differ->diff($yamlContent, $correctYamlContent);
 
-        return new Result($pathToYamlFile, Result::RESULT_CODE_INVALID_FILE_SYNTAX, $diffBetweenStrings);
+        return new Result($pathToYamlFile, Result::RESULT_CODE_INVALID_FILE_SYNTAX, ProcessOutput::STATUS_CODE_INVALID_FILE_SYNTAX, $diffBetweenStrings);
     }
 
     /**